requirements.txt
# Core ML and Data Processing
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Web Framework and API
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0

# Dashboard
streamlit==1.27.0
plotly==5.16.1
altair==5.1.2

# Blockchain and Web3
web3==6.11.1
requests==2.31.0
aiohttp==3.8.5

# Data Validation and Processing
pydantic==2.3.0
marshmallow==3.20.1

# Configuration and Environment
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development and Linting
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Docker and Containerization
docker==6.1.3

# Database (for audit trail)
sqlalchemy==2.0.20
alembic==1.12.0

# Security and Encryption
cryptography==41.0.4
passlib==1.7.4

# Additional Utilities
python-multipart==0.0.6
jinja2==3.1.2
markdown==3.5.1 